Fix: Nametag is not dynamic with wearable bounds (#1683) (#1950)
* fix: Nametag is not dynamic with wearable bounds (#1683)

It calculates the local bounds of the avatar when it is created and stores it in the AvatarCustomSkinningComponent. Whent he NamePlacementSystem uses that data to calculate the height of the nametag for each avatar.

* Added reference to Utility.Editor in DCL.Plugins

Uncommented a test code that shows the bounds of the avatars.

* Commented out the code that allows visualizing the bounds of the avatar

I left it there just in case it is useful for other programmer to check that bounds are correct.