feat: Add Playables to control Avatar (#1999)
Implemented a playable asset for the Timeline that can trigger desired emotes for a desired avatar, in editor.
Emotes can loop and will stop as soon as the clip ends.
In order to be able to access the ECS World, it was necessary to create an static property in the GlobalWorld class.
All this code will execute in Editor only.