feat: enforce latest version (#1954)
## What does this PR change?
Close #1900

Adds check of the current explorer version on startup and if not, it shows a UI popup where user can select or to Update the version or to close the Explorer without the update. The option with Update will start Launcher (if installed) or will start downloading the Launcher from download page and close Explorer. 

Feature is forced on the release and disabled for the debug mode.

## How to test the changes?

1. Launch the explorer from the command line with  `--simulateVersion 0.1.0  --debug --versionControl true`  parameter
2. After the splash screen you should see a popup
3. Press "Update Now" -> verify it opens Launcher and closes Explorer
4. Press "Exit Application" -> verify it just closes Explorer

Test on Win, Mac and Mac arm.
Test "Update Now" with Launcher installed and with Launcher not installed machines. 
Test "Update Now" with different Launcher installation folders (if supported).

Test with just parameters `--simulateVersion 0.1.0  --debug` -  it shouldn't  popup.