fix: Road assets are not being unloaded when jumping to a world (Memory leak) (#1992)
Now all road assets are destroyed when leaving Genesis plaza.
The Unload method of the RoadAssetPool was not being called ever because the frame time budget did not allow it. Apart from that, it would not destroy assets that were out of the pool, Clear requires elements be in the internal list to be destoryed.