fix: [QA] LOD Loads Only When Very Close to Scenes (#1728)
It makes LOD0 of "small" scenes to appear when the camera is farther than before.
Scenes with 10 parcels or less are considered "small" (arbitrary value, depending on performance). The smaller the scene is (under 11 parcels), the larger the distance from which the LOD0 is visible with respect to its original value.