Changes requested by the code review
TriggerEmotePlayableBehaviourcs.cs renamed to TriggerEmotePlayableBehaviour.cs.
Now all playable behaviours that animate an AvatarBase inherit from BaseAvatarPlayableBehaviour. It keeps the cached Entity and AvatarBase up to date and tells inherited classes when the avatar set in the track has changed.
Now the ResolveVelocity method of CalculateCharacterVelocitySystem has been refactored so its content has been moved to another method that receives a transform, and this transform depends on whether we are updating RandomAvatars or not (now there are 2 queries).