Feat: Add Playables to control Avatar (#1999) (#2079)
* feat: Add Playables to control Avatar (#1999)

Implemented a playable asset for the Timeline that can trigger desired emotes for a desired avatar, in editor.
Emotes can loop and will stop as soon as the clip ends.
In order to be able to access the ECS World, it was necessary to create an static property in the GlobalWorld class.
All this code will execute in Editor only.

* Implemented a workaround to play and download emotes that don't belong to the player

* Implemented playables for moving and teleporting avatars

Created the playable Move Avatar to let avatars walk forward and rotate.
Created the playable Teleport Avatar to set the position and rotation of avatars to specific values.
Existing systems modified in order to make random avatars independent, so their transformations are not affected by other features.
ASMdef name changed to DCL.AvatarAnimation.Editor.

* Fixed a problem related to the updating order of the playables

The Unity timeline updates in a sequence separated from the systems of the ECS world. That was causing that some data was overriding sometimes. Now a system is in charge of updating the timeline at the right moment.
Now then emote are stopped it also cleans emotes that are not loaded yet. Sometimes when loading an emote failed, it could not try to load other anymore.

* Fixed a namespace inclusion that should only be there in editor

* The system warns when PlayableDirector is not updated manually - Fixed a tiny bug

* Changes requested by the code review

TriggerEmotePlayableBehaviourcs.cs renamed to TriggerEmotePlayableBehaviour.cs.
Now all playable behaviours that animate an AvatarBase inherit from BaseAvatarPlayableBehaviour. It keeps the cached Entity and AvatarBase up to date and tells inherited classes when the avatar set in the track has changed.
Now the ResolveVelocity method of CalculateCharacterVelocitySystem has been refactored so its content has been moved to another method that receives a transform, and this transform depends on whether we are updating RandomAvatars or not (now there are 2 queries).

* Fixed: OnAvatarChanged was not being called

---------

Co-authored-by: Juan Ignacio Molteni <juanignaciomolteni@gmail.com>