fix: Emotes menu is overlapping the Wearables menu (#1981)
The E key, which brings you to the emote slots configuration menu, was enabled while the emotes wheel menu was not ready to be closed yet (the task that is awaiting for the controller to close is not listening to cancellations or completion yet). Now the key is enabled after the menu is ready.