Now it takes into account many corner cases
Calculations are performed only when the Transform of the TMP changes or when its properties change (we are only interested in text though).
It enables and disables the TMP in the VisibilityTextShapeSystem, according to what was calculated in the UpdateTextShapeSystem (added IsContainedInScene flag to TexhShapeComponent to transmit the result).
It properly calculates the bounds of the TMP when it is disabled by caching the latest valid bounds size (added vector to TextShapeComponent) and using it when the bounds are zero.
It takes into account the changes in the text of the TMP of the scene, and it calculates the visibility accordingly.
It forces the TMP to rebuild when the text changes, in order to be able to use the proper bounds.