fix: Unable to recover initial memory state in GP (#2129)
Now cached texture intentions are compared using the texture file hash instead of the URL that was used to download it. Now it does not matter which content server is the candidate to get the texture from, if the instance was previously stored, it will reuse it when entering Genesis plaza.