fix: The outline of text shapes is wrong when created
When the cached TextMeshPro instances were reused and the new outline width equaled zero, it tried to disable the outline shader keyword which does not exist in the shader (so the previous values related to outline stayed the same and were still used). Now it also resets the values of the related properties.