Video Audio Playback Prioritization (#2627)
Added a new slider to the environment settings for the maximum simultaneous videos that can be played at the same time.
Now videos are played or paused according to their position with respect to the player's camera. It takes into account the distance and the angle of the viewer with respect to the video mesh. This produces a score that puts every video that "should" be playing in a sorted list, from which only the first N videos are played and the rest are paused, where N = the maximum amount of videos allowed.
Videos that are behind the camera are not played.
Created the UpdateMediaPlayerVisibilitySystem.