Priority visualization helpers for debugging & Settings asset with weights
Added small colored cubes at the corner of each video screen so we can know the priority of that video (high priority: green, low priority: red, paused: blue, not prioritized: black).
Created the VideoPrioritizationSettings asset so it is easy to set the weights used for computing the priority of the videos.
Added the VideoPrioritizationSettings asset to the World Plugins Container.