fix: Videos were freezing if paused more than 10 secs and played again (#2782)
The problem was related to a mechanism to detect erroneous streams after no change in current time for more than 10 seconds. Now it checks that it was not manually paused previously and, if it was, then it plays normally. It also waits without error if the player was seeking or buffering.