Added distance and size limits to Video prioritization settings And refactoring
Now videos are discarded when they are too far or look too small on screen, according to 2 new settings parameters.
The calculation of the relative height of the meshes has changed.
Now the angle of the camera with respect to the videos uses flattened XZ vectors to avoid some big video meshes from pausing when the camera is too close to them (since currently the center of the mesh is used, a difference in height may produce a bigger angle so the video gets discarded as if it was at the sides of the camera).
Now debugging cubes are cleaned when the scene is unloaded.
Now VideoStateByPriorityComponents are removed when the MediaPlayers are removed in the clean up system.