Cached FOV & Screen size fix & Video size value changed
Now it uses the diagonal of the bounding box as size for the video.
The distance used to calculate the screen size of a video was wrong.
Now it caches the values of the vertical and horizontal FOV so priorities do not change while the character is running.
The values of the video prioritization settings asset have been tweaked.