Added support for meshes that share the same video player
The size of the group of videos is the distance from the minimum corner to the maximum corner of the sum of all the bounding boxes.
The center point of the group of videos is the middle point between the minimum corner and the maximum corner of the sum of all bounding boxes.