If camera is in the bounding box of a group of meshes that share 1 video, it isnt paused
Meshes that share the same video texture are grouped as if they were all one mesh. The bounding box of that group is used to compute whether the video has to play or not. Now it calculates whether the camera is contained in the bb, and if it is, the video will not be paused even if the center of the bb is behind the camera.