Video / Audio Playback Prioritization shape (#2812)
Now videos are played or paused according to their position with respect to the player's camera. It takes into account the distance and the angle of the viewer with respect to the video mesh. This produces a score that puts every video that "should" be playing in a sorted list, from which only the first N videos are played and the rest are paused, where N = the maximum amount of videos allowed.
Videos that are behind the camera are not played.
Created the UpdateMediaPlayerVisibilitySystem.
Added Cinemachine reference to MediaPlayer asdf.
Videos continue at current time instead of resuming at the time they were paused, like if they were never paused.
Now vertical size of the video meshes related to screen is calculated properly.
Videos outside of the horizontal FOV are not played.
Added small colored cubes at the corner of each video screen so we can know the priority of that video (high priority: green, low priority: red, paused: blue, not prioritized: black).
Created the VideoPrioritizationSettings asset so it is easy to set the weights used for computing the priority of the videos.
Added the VideoPrioritizationSettings asset to the World Plugins Container.
Added distance and size limits to Video prioritization settings And refactoring.
Now videos are discarded when they are too far or look too small on screen, according to 2 new settings parameters.
Now the angle of the camera with respect to the videos uses flattened XZ vectors to avoid some big video meshes from pausing when the camera is too close to them (since currently the center of the mesh is used, a difference in height may produce a bigger angle so the video gets discarded as if it was at the sides of the camera).
Debugging cubes are cleaned when the scene is unloaded.
VideoStateByPriorityComponents are removed when the MediaPlayers are removed in the clean up system.
It uses the diagonal of the bounding box as size for the video.
It caches the values of the vertical and horizontal FOV so priorities do not change while the character is running.
Global Plugins Settings asset modified to add the VideoPrioritizationSettings asset to QualityContainer.Settings and ExplorePanelPlugin.ExplorePanelSettings.
VideoPrioritizationSettings is now in a new asmdef DCL.MediaPlayer.Settings so it can be referenced by other modules without cycles. I has been added to other asmdefs.
Max simultaneous videos per quality level: Low -> 1, Medium -> 5, High -> 10.
The world position of the MediaPlayer is set when the component is created.
Modified the prefab of the MediaPlayer to make Spatial sound fully 3D and add a softer volume curve.
Added support for meshes that share the same video player
The size of the group of videos is the distance from the minimum corner to the maximum corner of the sum of all the bounding boxes.
The center point of the group of videos is the middle point between the minimum corner and the maximum corner of the sum of all bounding boxes.
Video prioritization settings asset added to Settings addressables group.
MediaPlayer transform is updated every frame if there is a TransformComponent present.
When there is a transform, the AudioSource is set as 3D. Otherwise it is set as 2D.
Meshes that share the same video texture are grouped as if they were all one mesh. The bounding box of that group is used to compute whether the video has to play or not. Now it calculates whether the camera is contained in the bb, and if it is, the video will not be paused even if the center of the bb is behind the camera.
MeshRenderer reference removed when a video texture consumer is no more.