fix: Sometimes videos that share the same texture do not start
The size of the videos were calculated once when the scene was loaded, but at that time not all videos were registered and properly placed. Therefore the size of a group of videos that use the same texture was wrong and that error stayed there. When calculating the priority of the group, it was too low and it was culled.
Now sizes are calculated in every iteration.