fix: Videos that play at a given time are starting at zero (#2955)
* fix: Videos that play at a given time are starting at zero

When the video was played, the MediaPlayerComponent.CurrentTime was read in order to know where to seek, but its value was zero sometimes (even if the MediaControl.Seek method was called previously). Now it reads IMediaControl.GetCurrentTime() instead.

* fix: Sometimes videos that share the same texture do not start

The size of the videos were calculated once when the scene was loaded, but at that time not all videos were registered and properly placed. Therefore the size of a group of videos that use the same texture was wrong and that error stayed there. When calculating the priority of the group, it was too low and it was culled.
Now sizes are calculated in every iteration.