ChatController and ChatView refactoring: Now the view is a black box for the controller
Many fields and methods moved from ChatController to ChatView and adapted so the code flow is the same but separating concerns. The controller does not know anything about how the view is implemented. ChatView now exposes some public methods for the controller to call them, as well as some events to which the controller has to subscribe.
Some dependencies injected to ChatController (emoji menu prefabs) are now directly referenced by ChatView in the Chat prefab.
Some setting constants exposed in the inspector of ChatView.
Now the chat bubbles toggle uses the isOn property and its events instead of showing / hiding images of the toggle manually.
Functionality has not changed.
The rest of scripts only contain renamed events.