ChatMessages do not store view data anymore, hasToAnimate removed
The flag hasToAnimate was added to the ChatMessage structure as a way to know which messages were already animated in the chat window (a fade-in FX is used when a new message arrives) when using LoopListView2, which reuses ChatEntryViews through internal pools. It made no sense to store data strictly related to a view in the model. This workaround is not necessary anymore, ChatView is now in charge of knowing how many messages to animate.