Implemented the concept of Chat channel, inside Chat history
Now it is possible to have multiple conversations that reuse the scroll view of the Chat view, and send messages to any of them.
Created the ChatChannelId struct which may be replaced by anything else in the future to identify a channel by its type and name.
Currently it's not possible to send messages to specific channels of other users, that has to be implemented in the future, although the base has been put in, the protobuf Chat message may store the Id of the channel at the start of the actual message and can be parsed by both endpoints (see MultiplayerChatMessageBus). Currently everything is sent to the special channel "nearby".
The ClearChatCommand had to be adapted since it has to be sent to the current chat channel, but only ChatView knows what's the current channel. So now the command does not change the Chat history, it calls the ChatController (injected in a later step during the initialization of the program).