Implemented the Main UI sidebar button that shows the number of unread messages
Created a new prefab variant for the new button. Added a new animation controller and all the animations of the icon.
Created a prefab for the numeric badge.
Added the new button at the bottom of the sidebar. The number updates as the amount of unread messages varies.
Connected the sidebar to the chat so it can open it when clicked.