Fix: Member list was scrolling back and forth when scrollbars were present
A VerticalLayout component in Content was causing that LoopListView was going crazy. Now the layout has been removed and the width of the items adjusted.
Added margins to Viewport, removed from Content (in Chat.prefab too).
Changed the fading effect in the viewport of the Member list, as top padding was removed (with the VerticalLayout), top is now zero.
The ChatMemberListItem prefab is now independent, it is not a variant of the Friends item prefab anymore. Unused components have been removed.