First working implementation
Removed the IChatLufecycleBusController from Chat and Friends.
Removed some shortcuts subscriptions from Chat and Friends.
Removed dependencies from PresistentFriendPanelOpenerController (MVCManager and DCLInput).
Implemented the Shared space manager, which is created in the container and sent to controllers. Each controller that competes for the space has to be registered by its plugin.
There is a fixed list of UI panels that can be shown / hidden, without knowing the implementation of any other controller.