Added the rest of panels to the shared space manager
Now opening any panel of the sidebar (except for the fullscreen views) hides the rest (the chat is hidden only when friends panel appears, otherwise, it just folds).
Created the IPanelInSharedSpace interface for any panel to be added to the manager.
Added the rest of the panels: Notifications, Skybox, EmotesWheel, SidebarSettings and SidebarProfile.
PersistentEmoteWheelOpenerController and ISidebarActionsBus can be removed.
Shared space manager refactored to reduce its code.