Many small changes in many places to make the manager rock solid
Added an event to the IPanelInSharedSpace interface. It is used to know when does the panel finishes showing.
Changed how panels are shown and hidden so they never overlap and don't break.
Cleaned and simplified the code in SidebarController.
Now the Emotes wheel panel does not disable shortcuts.
Now the Skybox button of the sidebar does not set interactable to false when clicked.
Now the Emotes wheel panel is not opened by the system but using the shortcut event.