Explore panel added to the shared space and MANY changes
Important: The State of the controller is set to ViewHidden after it has been hidden, not at the beginning of the call. Thanks to this, it is now possible to wait until it is actually invisible.
Now Friends and Emotes wheel are Popups instead of Persistent.
Now HideInSharedSpaceAsync make the controllers / MVC hide the view by themselves instead of calling HideViewAsync.
Chat panel visibility is handled in a special way. Its view is not hidden, as it is a persistent view, so it is made transparent or disabled.
Explore panel added to the shared space manager.
Shortcuts moved to the shared space manager.
Skybox panel does not use Overlay canvas ordering anymore.
In the Main UI, Friends panel overrides the canvas sort order, otherwise the popup closer appears over it instead of beneath of it.
Everything working as intended, although InWorldCamera panel is still pending, and the Chat should receive a parameter that tells when to get focused.