The Profile menu controller now behaves as the other popups
However, when used in the Backpack window button, it works in a different way because the MVC manager does not allow more than one instance of a controller (as each controller is registered by its type).
The sort order of the ProfileMenuView prefab is now overridden, to make the popup closer work properly.