Fix: Meta-GamiMall platform collider dissapears (#3499)
# Pull Request Description

Collider on the platform was still out of scene bounds in some places and get disabled, which prevent platform for going further.

## What does this PR change?
I just increased offset and add a check so bounds are not inverted in case when they are smaller than the offset.

## Test Instructions
1. Go to Meta GamiMall `/goto 147,60`
2. Jump on the shuttle pad and start going through the track
3. Do the full ride (it last more than 2 minutes) - verify you can always stay on the platform and finish the ride 

### Additional Testing Notes
Note: issue with not able to stand in the center of the platform is out of the scope of the testing. Main test scope here is that you can finish the ride.