Inworld camera shortcut moved to the SharedSpaceManager
Now it waits for the other panels to hide before showing the camera.
Added ECS, ECS.Unity and ScreenCapture to SharedSpaceManager asmdef.
This fixes the problem of the panels not opening anymore after the in-world camera was shown and the issue related to not showing the chat if the Friends panel was open.