Fix: Meta GamiMall | player pushed out of platform (#3556)
# Pull Request Description
Fix #3378 

FPS drops were causing issue of player being pushed from the platform. It was because application of platform delta between these 2 frames was done via characterController.Move

## What does this PR change?
- similarly to web explorer, changed application of platform delta from `characterController.Move` to direct application to the position

## Test Instructions
To easily test it you need to create FPS drop. This can be achieved by repositioning application in the OS when it is in Windowed mode (dragging it by the top bar)

### Test Steps
1. Go to Meta GamiMall  - `/goto 147,60`
2. Start the pad and stay on it while it is going through the track
3. Make FPS drop 
4. Verify that you avatar is always presented on the pad and FPS lags are not pushing it away from it