Fix: roads distance being much father than in non GPU-instanced approach (#3658)
# Pull Request Description
When we presented GPU Instanced, we set rendering distance to be dependent on the Environment distance. However, previously roads rendering distance was depending on Scene Distance. This caused not clear data about performance change with presenting this feature. 
In this PR we still use Environment Distance, but scale it approximately to values of Scene Distance.

## What does this PR change?
- Add scaling for rendering distance for GPU instanced renderering
- Add debug widget to be able enable/disable GPU Instancing and also adjust its scale factor

## Test Instructions

1. Go to AirDrop tower to see a roads from the top
2. In the GPU Instancing debug widget switch on/off "Is Enabled" toggle -> verify Roads are visible when it is enabled and not visible when disabled.
3. Set "Is Enabled" to true;
4. Go to Settings and set Env Distance to maximum (7000)
5. Play with "EnvDist ScaleFactor" in Debug window for GPU Instancing and see that it how far roads are rendered. 

Additionally, compare how far roads are rendered with this screenshots from v 0.57. release of the client (or make you screenshots). It should be approximately near

Scene Distance 20 ( equal to Environment Distance 1000 in this PR)
https://reels.decentraland.org/5341bacf-3fd4-411d-ace9-6e4a1e93373a

Scene Distance 60 ( equal to Environment Distance 4000 in this PR)
https://reels.decentraland.org/d2700aff-0adb-4e96-8ab6-5b26d8d90cb0

Scene Distance 100 ( equal to Environment Distance 7000 in this PR)
https://reels.decentraland.org/3d08cd41-582e-4862-8492-25a4c8d6d769

Real Roads on these screenshots are these that looks brighter.

### Additional Testing Notes
- Use InWorld Camera zoom to check rendering in distance
- Distance should match approximately, not exactly
- In new approach (this PR), you will see more details in the distance compared to what was in the 0.57 release (bushes, lamps and benches are not culled). It is known behavior and not a part of this PR.