The layout of the camera does not appear when triggering the C key multiple times (#3752)
Now the InWorldCameraController ignores requests to be shown if it is already visible, and ignores requests to hide if it is not visible.
Now the flow of closing is the same when pressing the close button and when triggering the C key shortcut, it now leaves the WaitForCloseIntentAsync method, which was not being left when the shortcut was used to hide the controller.
Fixed also: Now InWorldCameraController is not hidden and shown when the screen is capture, instead the view is just disabled and enabled.