Fix: TrackPlayersPositionSystem preformance
# Pull Request Description

As at can be see from the screenshot, `TrackPlayersPositionSystem` was spending time for  `WaitForJobGroupID`  of `BoneMatrixCalculationJob`.

 There is no any direct connection between these two systems. Most likely, it is related on how Jobs that use `TransformAccessArray` work (which in our case is `BoneMatrixCalculationJob`). 
As mentioned in this [discussion](https://discussions.unity.com/t/debugging-why-unity-urp-is-forcing-long-running-job-to-complete-on-main-thread-with-waitforjobgroupid/1618021)
> If you’re using `TransformAccessArray `in your jobs, any code that touches these transforms might cause the jobs to complete.

As a solution I moved execution of `TrackPlayersPositionSystem` to a later stage (to `PreRenderingGroup`), so job has more time to be completed.

I informed @dalkia about the issue with that job and regarding its possible re-work in future:
> If it’s a long-running job, it might be better to read the transform data once at the start (in a separate job?) so that the processing isn’t dependent on the scene.

## Test Instructions
It is tiny performance optimization that affects only Markers on the map and mini-map. So just verify that these player markers are rendered and corresponds to players positions. It only affects other players markers, not your avatar.