All functionalities implemented
Now encryption and decryption work, both for the name of the files and for their content.
Now it stores and loads the conversation settings.
Now it is possible to close a conversation.
Now stored history is loaded when opening a private conversation that was previously closed.
Now when a message is received for a conversation that was closed or never opened, it stores the file.
Now the name of the user is stored for every chat message.
Implemented the ChatMessageFactory so we can create messages in the same way from different places.
Now the nearby channel is open when a private conversation is closed.
Now only the chat conversations that were open in the toolbar when leaving the application are open when entering again. Closed conversations are not listed, even if their files exist. If the file that stores this info is not present, then all stored conversations are visible in the toolbar.