Now Chat controller and view are prepared for user log in and out
Logic in ChatController moved from OnViewInstantiated to OnViewShow so initialization code can occur more than once (the view is never destroyed).
Now the ChatController can be shown and hidden. This will occur when the user logs in or logs out respectively. It is the plugin who will call them so the controller does not know when it has to appear or disappear.
When the ChatController is hidden, it stops listening or reacting to any other system.
Event unsubscriptions moved from Dispose to HideAsync / OnViewClose.
Fixed a problem related to not being able to set the encryption key when the user was not logged in from the beginning.