Fix: The HaskKey class produced different hashes for same input
It internally uses the ArrayPool.Rent method which may (or may not) return a byte array that is longer than the amount asked, according to Microsoft's doc, so when that array is filled with the bytes of the string for which we want to calculate the SHA256 hash, it is passing the entire array to the SHA hashing function (with unknown length, and unknown content in not written bytes at the end of the array), which may produce a different hash for the same input string in different moments of the execution.
PS: The only way to reproduce the bug was using a build, not in editor.