Local storage of the chat history (private conversations shape) (#3827)
* First upload, main functionality already set, not implemented (#3707)

* All functionalities implemented

Now encryption and decryption work, both for the name of the files and for their content.
Now it stores and loads the conversation settings.
Now it is possible to close a conversation.
Now stored history is loaded when opening a private conversation that was previously closed.
Now when a message is received for a conversation that was closed or never opened, it stores the file.
Now the name of the user is stored for every chat message.
Implemented the ChatMessageFactory so we can create messages in the same way from different places.
Now the nearby channel is open when a private conversation is closed.
Now only the chat conversations that were open in the toolbar when leaving the application are open when entering again. Closed conversations are not listed, even if their files exist. If the file that stores this info is not present, then all stored conversations are visible in the toolbar.

* Refactor moving functionality to new classes

Encryption-related functionality moved to ChatHistoryEncryptor.
Serialization-related functionality moved to ChatHistorySerializer.
Added a mutex enveloping the usage of the channels dictionary.

* Prepared the code for when a different user logs in

* Minor changes

* Fixes after merging and optimization of ChatMessage

Removed IsPrivateMessage and ChannelId from ChatMessage.

* Added support for feature flags

* All new classes documented

* Folders were not updated when encryption key changed

Unified the flow for changes in the encryption key.
Other small refactor changes here and there. The initialized channels will ignore initialization calls. Clear was not well implemented.

* Minor changes

* ChatStorage renamed to ChatHistoryStorage

* Json serialization moved to ChatHistorySerializer + Fixes from review

* Fixed: The unread messages were displayed when entering a channel

* Now Chat controller and view are prepared for user log in and out

Logic in ChatController moved from OnViewInstantiated to OnViewShow so initialization code can occur more than once (the view is never destroyed).
Now the ChatController can be shown and hidden. This will occur when the user logs in or logs out respectively. It is the plugin who will call them so the controller does not know when it has to appear or disappear.
When the ChatController is hidden, it stops listening or reacting to any other system.
Event unsubscriptions moved from Dispose to HideAsync / OnViewClose.
Fixed a problem related to not being able to set the encryption key when the user was not logged in from the beginning.

* Added extra logging for debugging

CHAT_HISTORY logs enabled.

* More logs

* Revert this: Disabled conditionals in ReportHub

* Fixed a problem with the chosen field separator

When using commas in the messages, the parser was treating that chat as a separator.

* More logs

* More logs

* More logs

* Trying with SHA256 encryptor directly

* Logs removed, Reporthub reverted

Error and Exception logging enabled in both prod and dev for CHAT_HISTORY.

* Fixed: The user folder was not being created the first time

When adding a channel without messages the first time, it was not stored.

* Fixed: The length of the encrypted stream was calculated wrong

* Enabling extra logs