Inputbox is always focused when chat panel is focused
Now the chat panel appears unfocused by default.
Now the input box has always the focus while the chat panels has it.
Now opening a context menu of the chat does not make it lose the focus.
Now the chat has the focus when opened from the sidebar and by pressing the T shortcut.
Now when pressing Enter while the inputbox is empty, does not hide the caret.
Added IsFocused, Focus and Blur methods to ChatView.