Many fixes on the behaviour of the chat UI elements!
Now the chat viewer's scrollbar hides along with the background.
Now the chat panel folds properly and the input box stays without focus.
Added the concept of "focused" to the input box.
Now when click event occurs, a ray is casted in the chat panel to see if it occurred outside of the panel or not, and ignores the close button.
The emojis and suggestion panels are being opened and closed properly. Clicking on them makes the chat panel gain the focus.
Now when closing the Member list the input box is focused.
Now when clicking on the toolbar, the input box is focused.
Now opening or closing any context menu does not affect the focus of the chat panel.