fix: restore scroll functionality and correct platform-specific sensitivity
Sensitivity for windows is increased.

NOTE: Unity issue when enabled image on the scroll view for chat to enable scrolling to work properly, sets the chat panel canvas scale to zero and it needs manually to be set again to 1. Probably happening because of deep nested prefabs or layout groups.