Community conversation: New chat panel view (#4178)
Added a new prefab to show the title of a community conversation.
Added a new prefab to show the conversation icon in the chat conversation toolbar.
Now adding a channel has been split into 3 to allow different type of data to be provided depending on the type of conversation.
IProfileThumbnailCache renamed to IThumbnailCache, as an attempt to make it more generic. IThumbnailCache is used for getting the image of the community asynchronously.
Now the multiplayer bus handles community conversation messages too.
Now community messages send a topic along with the messages. Messages are sent to a special user created server-side which resends them to members of the community specified in the topic.
Now a new conversation is added to the chat when the user logs in for each community he belongs to.
Multiple users can communicate using a community conversation.